Still more changes on UnitSelection code
git-svn-id: https://mary.opendfki.de/repos/trunk@93 953a6561-930b-0410-b2a6-db37d1b2ae63

+++ java/de/dfki/lt/mary/unitselection/ZeroTargetCostFunction.java
@@ -47,7 +47,7 @@ public class ZeroTargetCostFunction implements TargetCostFunction
     {
     }
 
-    public int cost(Target target, Unit unit)
+    public float cost(Target target, Unit unit)
     {
         return 0;
     }
