+++ java/de/dfki/lt/mary/unitselection/viterbi/ViterbiPath.java
@@ -38,7 +38,7 @@ import de.dfki.lt.mary.unitselection.viterbi.ViterbiCandidate;
   * Describes a Viterbi path.
   */
  public class ViterbiPath {
-	private int score = 0;
+	private float score = 0;
 	private ViterbiCandidate candidate = null;
 	private LinkedHashMap f = null;
 	private ViterbiPath previous = null;
@@ -48,7 +48,7 @@ import de.dfki.lt.mary.unitselection.viterbi.ViterbiCandidate;
 	 * Get the score of this path
 	 * @return the score
 	 */
-	public int getScore(){
+	public float getScore(){
 	    return score;
 	}
 	
@@ -56,7 +56,7 @@ import de.dfki.lt.mary.unitselection.viterbi.ViterbiCandidate;
 	 * Set the score of this path
 	 * @param score the new score
 	 */
-	public void setScore(int score){
+	public void setScore(float score){
 	    this.score = score;
 	}
